requirements.txt
streamlit==1.38.0
langchain==0.2.16
langgraph==0.2.17

sentence-transformers==3.0.1
faiss-cpu==1.12.0         # use this newer wheel

yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4             # stable wheel on py3.11+arm64; plenty for this app
pymupdf==1.24.9
pdfplumber==0.11.4
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.32
